Break apart TileEngine Actor Orders into class files
Having all the Orders declared in a single file was kind of a lazy way
to keep them all grouped together. Directory structure can be used to
do the same thing, and the end result is that the orders can be split
up into their own class files, consistent with the style of the rest of
the codebase, while still being logically grouped together.
